Moved methods to utils
