+++ android-wsbridge/app/src/main/java/com/dji/wsbridge/lib/Utils.java
@@ -1,6 +1,7 @@
 package com.dji.wsbridge.lib;
 
 import com.dji.wsbridge.BuildConfig;
+import com.google.firebase.crashlytics.FirebaseCrashlytics;
 
 import java.io.BufferedInputStream;
 import java.io.ByteArrayOutputStream;
@@ -186,5 +187,17 @@ public class Utils {
 
     public static boolean isInternalVersion() {
         return BuildConfig.BUILD_TYPE == "internal";
+    }
+
+    public static void logToFirebase(String message) {
+        if (isInternalVersion() && FirebaseCrashlytics.getInstance() != null) {
+            FirebaseCrashlytics.getInstance().log(message);
+        }
+    }
+
+    public static void recordExceptionToFirebase(Exception exception) {
+        if (isInternalVersion() && FirebaseCrashlytics.getInstance() != null) {
+            FirebaseCrashlytics.getInstance().recordException(exception);
+        }
     }
 }
